requirements.txt
# requirements.txt

# Core
fastapi
uvicorn[standard]

# API & Data
openai
httpx
pydantic
pydantic-settings
python-dotenv
groq  # <-- ADD THIS LINE

# File Handling & OCR
python-multipart
pytesseract
pillow
python-docx
PyMuPDF
pdf2image

# PDF Generation
fpdf2